refactor code by removing util.go
